feat(hiddenColumnsSearchMessage): Moved filterTableData to new file
