+++ src/Analysis/AtlasDataDictionary/AtlasDataDictionaryTable/Components/SearchBar/SearchBar.tsx
@@ -4,6 +4,7 @@ import { debounce } from 'lodash';
 import SearchIcon from '../Icons/SearchIcon';
 import { IRowData } from '../../Interfaces/Interfaces';
 import SearchBarMessage from './SearchBarMessage/SearchBarMessage';
+import filterTableData from './filterTableData';
 
 interface ISearchBarProps {
   columnsShown: number;
@@ -22,38 +23,10 @@ const SearchBar = ({
   handleTableChange,
   columnManagementReset,
 }: ISearchBarProps) => {
-  const filterTableData = () => {
-    const filteredData = TableData.filter((item) => {
-      const searchQuery = searchTerm.toLowerCase().trim();
-      return Object.values(item).some((value) => {
-        if (typeof value === 'string' || typeof value === 'number') {
-          return value.toString().toLowerCase().includes(searchQuery);
-        }
-        if (Array.isArray(value)) {
-          let doesArrayContainsSearchQuery = false;
-          value.forEach((arrItem) => {
-            Object.values(arrItem).some((arrObjValue) => {
-              if (
-                (typeof arrObjValue === 'string'
-                  || typeof arrObjValue === 'number')
-                && arrObjValue.toString().toLowerCase().includes(searchQuery)
-              ) {
-                doesArrayContainsSearchQuery = true;
-              }
-              return null;
-            });
-          });
-          return doesArrayContainsSearchQuery;
-        }
-        return null;
-      });
-    });
-    setData(filteredData);
-  };
   const debounceDelayInMilliseconds = 500;
   useEffect(() => {
     const debouncedFilter = debounce(() => {
-      filterTableData();
+      filterTableData(TableData, searchTerm, setData);
     }, debounceDelayInMilliseconds);
     debouncedFilter();
   }, [searchTerm]);
@@ -61,6 +34,8 @@ const SearchBar = ({
   const handleInputChange = (event) => {
     handleTableChange('searchTerm', event.target.value);
   };
+
+  const showSearchBarMessage = true;
 
   return (
     <thead className={'search-bar'} data-testid='search-bar'>
@@ -92,10 +67,11 @@ const SearchBar = ({
                 onChange={(e) => handleInputChange(e)}
               />
             </div>
-            <div className='column'>
-              <SearchBarMessage columnManagementReset={columnManagementReset} />
-            </div>
-
+            {showSearchBarMessage && (
+              <div className='column'>
+                <SearchBarMessage columnManagementReset={columnManagementReset} />
+              </div>
+            )}
           </div>
         </th>
       </tr>
